Fix compilation errors and compilation warnings on MSVC++.

+++ src/Module/Encoder/RSC/Encoder_RSC_generic_sys.cpp
@@ -8,7 +8,7 @@ template <typename B>
 Encoder_RSC_generic_sys<B>
 ::Encoder_RSC_generic_sys(const int& K, const int& N, const int& n_frames, const bool buffered_encoding,
                           const std::vector<int> poly, const std::string name)
-: Encoder_RSC_sys<B>(K, N, std::floor(std::log2(poly[0])), n_frames, buffered_encoding, name),
+: Encoder_RSC_sys<B>(K, N, (int)std::floor(std::log2(poly[0])), n_frames, buffered_encoding, name),
   out_parity(),
   next_state(),
   sys_tail  ()
